backend/requirements.txt
fastapi
pytz # Timezone handling for reminders
uvicorn[standard]  # Uvicorn with standard dependencies (needed for production)
sqlalchemy
psycopg2-binary  # PostgreSQL database adapter
pydantic
passlib[bcrypt]  # Secure password hashing
python-jose[cryptography]  # JWT handling for authentication
python-dotenv  # For managing environment variables
httpx  # Async HTTP requests
ics  # iCalendar file handling
icalendar
email-validator  # Email validation
python-multipart # files download/upload between program and user
requests  # HTTP requests
pytest  # Testing framework
google-genai # Google Natural Language API
redis # redis for caching